feat: add style-dictionary tokens for download
